added the files available on the site.
if you have something to change, feel free to modify them. When 
you feel it's time to sync the site, if you have write access to 
the site, feel free to do it (look at the upload_site.sh script
to see how i do it). if not, just send me a message and I can do this
for you.