Update for jsportal
Now the script can draw secondary menu links. This is used on the docs menu.

Other changes:
 - The iframe has no border. Now it will not be seen on Firefox
 - The script for generating the menus has been removed from the onload.
   Now the internal content might be seen before the page is fully loaded (when waiting for 
   external images, etc)
 - The main menu has separators.
 - Added a new section which talks about release versions.
 
The new site is not loadaded to the site yet, as I am waiting your approval. Meanwhile, you can test
it on your local machine and in this url:
http://devqt.berlios.de/tmp/test/

The URL will be removed as soon as this site is uploade