added a pre release of the plugin demo.
This demo shows how to code a plugin system. Each plugin is self
contained and does not have to know anything about the main window, or
the system used for the mdi. All the comunication is done via qmdiServer
and qmdiHost.

The application has also a configuration dialog, which will be (not
implemented yet) with all the plugins from which you can enable/disable
the available plugins (unless always enabled is on). The configuration
dialog will also contain several pages for each plugin, if that plugin
contains a configuration widget (not implemented yet). Each wdget will
also have a configuration item prepared from which it will load/save
it's
state it on load/unload (not implemented yet).

the demo currently abuses the code from the mdi-tab demo: it contains 2
plugins : for file editing and for help. The plugins display the same
widgets at the tab demo. The functionality of this application is not as
good as the original demo, but it's on the way.

todo:
 1) Plugins are hard coded/linked into the application. Need to find a
    way to make the plugins loadable at run time. This means that you
    can theoratically edit the plugins from the main application and
    while it's working just unload/load the plugins to get newer
    versions.
 2) Need to add new functionalities to mdi-clients: "can you close
    yourself?" This will be used for providing a way to query the user
    if he wants to save his work before closing the mdi-client.
 3) QSA. Programmers should have the ability to write plugins in JS/QSA
    scripts.
 4) What if the plugins need plugins? for example, used defined plugins
    for the editor.
 5) Interaction between plugins: how can the help plugin extend the
    editor plugin?
 6) Dependencies between plugins: a project plugin will need the editor
    plugin to be loaded.
 7) Clean up the code. It's a mess.
 8) Documentation.